+++ ndex2cx/nice_cx_builder.py
@@ -1,9 +1,6 @@
 __author__ = 'aarongary'
 
 import logging
-import json
-import ijson
-import requests
 import base64
 import sys
 import math
@@ -59,7 +56,8 @@ class NiceCXBuilder(object):
 
     def set_context(self, context):
         """
-        Set the @context information of the network.  This information maps namespace prefixes to their defining URIs
+        Set the @context information of the network.  This information
+        maps namespace prefixes to their defining URIs
 
         Example:
 
@@ -101,7 +99,6 @@ class NiceCXBuilder(object):
         :return: None
         :rtype: None
         """
-
         add_this_network_attribute = {'n': name, 'v': values}
         if type:
             add_this_network_attribute['d'] = type
@@ -208,7 +205,6 @@ class NiceCXBuilder(object):
             raise TypeError('Attribute value is None')
 
         add_this_node_attribute = {'po': property_of, 'n': name, 'v': values}
-
 
         if self.node_attribute_map.get(property_of) is None:
             self.node_attribute_map[property_of] = {}
@@ -440,44 +436,6 @@ class NiceCXBuilder(object):
                     return_list.append(aspect[key])
 
         return return_list
-
-    def load_aspect(self, aspect_name):
-        #with open('Signal1.cx', mode='r') as cx_f:
-        with open('network1.cx', mode='r') as cx_f:
-            aspect_json = json.loads(cx_f.read())
-            for aspect in aspect_json:
-                if aspect.get(aspect_name) is not None:
-                    return aspect.get(aspect_name)
-
-    def stream_all_aspects(self, uuid):
-        return ijson.items(urlopen('http://dev2.ndexbio.org/v2/network/' + uuid), 'item')
-
-    def stream_aspect(self, uuid, aspect_name):
-        if aspect_name == 'metaData':
-            self.logger.debug('http://dev2.ndexbio.org/v2/network/' + str(uuid) + '/aspect')
-            md_response = requests.get('http://dev2.ndexbio.org/v2/network/' + uuid + '/aspect')
-            json_respone = md_response.json()
-            return json_respone.get('metaData')
-        else:
-            request = Request('http://dev2.ndexbio.org/v2/network/' + uuid + '/aspect/' + aspect_name)
-            base64string = base64.b64encode('%s:%s' % ('scratch', 'scratch'))
-            request.add_header("Authorization", "Basic %s" % base64string)
-            #result = urllib2.urlopen(request)
-            urlopen_result = None
-            try:
-                urlopen_result = urlopen(request) #'http://dev2.ndexbio.org/v2/network/' + uuid + '/aspect/' + aspect_name)
-            except HTTPError as e:
-                self.logger.error(str(e.code))
-                return []
-            except URLError as e:
-                self.logger.error('Other error')
-                self.logger.error('URL Error ' + str(e.message()))
-                return []
-
-            return ijson.items(urlopen_result, 'item')
-
-    def stream_aspect_raw(self, uuid, aspect_name):
-        return ijson.parse(urlopen('http://dev2.ndexbio.org/v2/network/' + uuid + '/aspect/' + aspect_name))
 
     def _infer_data_type(self, val, split_string=False):
         """